reorg site-license related types and add idle timeout
